continued work on new webfrontend admin processes
