+++ web/tomato/templates/organization/info.html
@@ -57,7 +57,7 @@
   <a class="btn btn-info" href="javascript:window.open('{%url "admin.organization.usage" organization.name%}', '_blank', 'innerHeight=450,innerWidth=600,status=no,toolbar=no,menubar=no,location=no,hotkeys=no,scrollbars=no');"><span class="glyphicon glyphicon glyphicon-stats"></span> Resource Usage</a>
   {% if request.session.user and request.session.user|args:organization.name|call:"isHostManager" %}
 	<div class="skip-xs"></div>
-	<button type="submit" formaction="{%url "site.add" organization.name%}" class="btn btn-success"><span class="glyphicon glyphicon-plus"></span> Add Site</button>
+	<button type="submit" formaction="{%url "admin.site.add" organization.name%}" class="btn btn-success"><span class="glyphicon glyphicon-plus"></span> Add Site</button>
 	<button type="submit" formaction="{%url "admin.organization.edit" organization.name%}" class="btn btn-warning"><span class="glyphicon glyphicon-pencil"></span> Edit Organization</button>
 	<button type="submit" formaction="{%url "admin.organization.remove" organization.name%}" class="btn btn-danger"><span class="glyphicon glyphicon-trash"></span> Remove Organization</button>
   {% endif %}
@@ -106,8 +106,8 @@
   <button type="submit" formaction="{%url "site_hosts" site.name%}" class="btn btn-info"><span class="glyphicon glyphicon-list"></span> Hosts</button>
   <button type="submit" formaction="{%url "site_external_networks" site.name%}" class="btn btn-info"><span class="glyphicon glyphicon-list"></span> External Network Instances</button>
   {% if request.session.user and request.session.user|args:organization.name|call:"isHostManager" %}
-  <button type="submit" formaction="{%url "site.edit" site.name%}" class="btn btn-warning"><span class="glyphicon glyphicon-pencil"></span> Edit Site</button>
-  <button type="submit" formaction="{%url "site.remove" site.name%}" class="btn btn-danger"><span class="glyphicon glyphicon-trash"></span> Remove Site</button>
+  <button type="submit" formaction="{%url "admin.site.edit" site.name%}" class="btn btn-warning"><span class="glyphicon glyphicon-pencil"></span> Edit Site</button>
+  <button type="submit" formaction="{%url "admin.site.remove" site.name%}" class="btn btn-danger"><span class="glyphicon glyphicon-trash"></span> Remove Site</button>
   {% endif %}
 </form>
 {% endfor %}
