+++ web/tomato/urls.py
@@ -68,10 +68,10 @@ urlpatterns = patterns('',
 	(r'^element/(?P<id>\d+)/console$', 'tomato.element.console'),
 	(r'^element/(?P<id>\d+)/console_novnc$', 'tomato.element.console_novnc'),
 	(r'^statistics$', 'tomato.main.statistics'),
-	(r'^map/$', 'tomato.site.map'),
-	(r'^map.kml$', 'tomato.site.map_kml'),
-	(r'^link_stats/(?P<site>\w+)$', 'tomato.site.details_site'),
-	(r'^link_stats/(?P<src>\w+)/(?P<dst>\w+)$', 'tomato.site.details_link'),
+	(r'^map/$', 'tomato.site_map.map'),
+	(r'^map.kml$', 'tomato.site_map.map_kml'),
+	(r'^link_stats/(?P<site>\w+)$', 'tomato.site_map.details_site'),
+	(r'^link_stats/(?P<src>\w+)/(?P<dst>\w+)$', 'tomato.site_map.details_link'),
 	url(r'^host/$', 'tomato.host.list', {"organization": None, "site": None}, name="host_list"),
 	url(r'^organization/(?P<organization>\w+)/hosts$', 'tomato.host.list', name="organization_hosts"),
 	url(r'^site/(?P<site>\w+)/hosts$', 'tomato.host.list', name="site_hosts"),
@@ -87,9 +87,12 @@ urlpatterns = patterns('',
 	(r'^organization/(?P<name>\w+)/edit$', 'tomato.admin.organization.edit'),
 	(r'^organization/(?P<name>\w+)/remove$', 'tomato.admin.organization.remove'),
 	(r'^organization/(?P<name>\w+)/usage$', 'tomato.admin.organization.usage'),
-	(r'^organization/(?P<organization>\w+)/add_site$', 'tomato.site.add'),
-	(r'^site/(?P<name>\w+)/edit$', 'tomato.site.edit'),
-	(r'^site/(?P<name>\w+)/remove$', 'tomato.site.remove'),
+	(r'^organization/(?P<organization>\w+)/add_site$', 'tomato.admin.site.add'),
+    (r'^site/$', 'tomato.admin.site.list'),
+    (r'^site/add$', 'tomato.admin.site.add'),
+    (r'^site/(?P<name>\w+)/edit$', 'tomato.admin.site.edit'),
+    (r'^site/(?P<name>\w+)/info$', 'tomato.admin.site.info'),
+	(r'^site/(?P<name>\w+)/remove$', 'tomato.admin.site.remove'),
 	url(r'^template/$', 'tomato.template.list', {"tech": None}, name="template_list"),
 	url(r'^template/bytech/(?P<tech>\w+)$', 'tomato.template.list', name="template_list_bytech"),
 	(r'^template/add$', 'tomato.template.add'),
